Fixed another geo circle/rect relate bug.
Couldn't use the proportions technique because of non-linear skew as you go north & south. If the rect crosses the x axis of the circle then it's necessary to check both corners because we can't reliably detect which is farthest.  Any way, the new logic is quite a bit clearer and still calls expensive contains() only 1.5 times or so (once or twice).