[CRS] Refactoring normalizePoint within a generic CRS package
To begin laying the foundation to support multiple projections a crs package is created. This package defines an extendable CRSDelegate that is instantiated by the SpatialContextFactory. The delegate is responsible for handling the logic defined by the coordinate reference system. The delegate also serves as the primary integration point for Proj4J, decoupling the transformation logic from the rest of the shape logic.