Moved DatelineRule from JtsWKTReader to top level.
The dateline rule set on the spatial contedxt factory is now used to configure
dateline handling when calling the non-explicit version of makeShape(). This
change also simplifies the calls to makeShape from JtsWktReader, rather than
check the dateline rule explicitly it can rely on the value set on the spatial
context.

Signed-off-by: jdeolive <jdeolive@gmail.com>